+++ src/adapters/postgresql/schemas/index.ts
@@ -397,6 +397,8 @@ export {
   KcacheQueryStatsSchemaBase,
   KcacheQueryStatsSchema,
   KcacheTopConsumersSchema,
+  KcacheTopCpuSchemaBase,
+  KcacheTopIoSchemaBase,
   KcacheDatabaseStatsSchemaBase,
   KcacheDatabaseStatsSchema,
   KcacheResourceAnalysisSchemaBase,